Asyncify prefetch_content code
Summary:
This diff may not have quite the right semantics.

It switches `prefetch_content` to async syntax,
in the process getting rid of the old function `spawn_future`,
which assumes old-style futures, in favor of using
`try_for_each_concurrent` to handle concurrency.

In the process, we were able to remove a couple levels of clones.

I *think* that the old code - in which each call to `spawn_future`
would spin off its own future on the side but then also wait
for completion, and then we buffered - would run at most 256
versions of `prefetch_content_node` at a time, and the current
code is the same. But it's possible that I've either halved or
doubled the concurrency somehow here, if I lost track of the
details.

Reviewed By: krallin

Differential Revision: D20665559

fbshipit-source-id: d95d50093f7a9ea5a04c835baea66e07a7090d14