mononoke: do not cache gets in MemWritesBlobstore
Summary:
At the moment MemWritesBlobstore stores both writes and reads. This is not
always desirable - in particular, in the next diff I'd like to add --dry-run
mode to backfill_derived_data which would keep everything that was derived in
memory. In that case storing all data that we read from blobstore might be too
much.
Given that we don't need this functionality let's just remove it

While there also add a method to get access to the underlying cache - again, it
will be used in the next diff.

Reviewed By: krallin

Differential Revision: D21088794

fbshipit-source-id: 91c8729d748d8ad8d9a70e6f8d5e15afe5021e8c