mononoke: [RFC] introduce generic Large and Small newtype wrappers
Summary:
To be used in the next diff. The idea is to provide a little more type safety at least on the function call boundary. We still have to access the wrapped type every so often, since we need to run the actual business logic on it. It seems still safer this way though.

Note: if this and the following diff look ok to people, I want to go over the actual commit sync lib and do the same there, add `Source` and `Target` wrappers, define wrapper-sensitive `Mover` and `BookmarkRenamer` types and so on.

Reviewed By: StanislavGlebik

Differential Revision: D20463054

fbshipit-source-id: 9049cd05327ee203f94df4e493e20b2615e617f1