scs_server: add commit_history to return the history of commits
Summary:
Add the `commit_history` method to the source control service.

This methods returns the history of the commit, directly from the changelog.

This differs from `commit_path_history` in that it includes all commits, whereas `commit_path_history` will skip commits that do not contain changes that affect the path.

Differential Revision: D21201705

fbshipit-source-id: dbf1f446c106620620343122176eccd5d809779c