mutationstore: add add_entries method
Summary:
This commit adds the `add_entries` method to the mutation store, which
allows Mononoke to add new entries to the store for a given set of
commits.

It is expected that the client will provide all of the mutation entries related
to the commits it is sending.  This may be too much information, as in normal
operation we will already know about the predecessors.  In that case we can
avoid additional work by just storing the entries directly related to the new
commits.

If the client has made commits while offline, then some mutation entries may
refer to predecessors that are not known locally.  In this case we must search
through the mutation history looking for the last commits we did know about,
and add all the in-between entries, too.

Reviewed By: StanislavGlebik

Differential Revision: D20287383

fbshipit-source-id: e5fb42bc4da7873c3a5aafd83684d374d9155bca