Remove CurrentState from EdenDispatcher
Summary:
The new merged EdenMount uses the sqlite overlay to track the status. Removing the calls to CurrentStatus from EdenMount.

This change is required here before we can switch to the merged EdenMount. Later in this stack we will add calls to the Inode APIs to report the FS changes.

Reviewed By: simpkins

Differential Revision: D20895074

fbshipit-source-id: 5e04684203f5735ff9068d3609b256991d1569f4