lfs_protocol: move LFS MIME type into protocol crate
Summary: In the process of factoring out generally useful parts of the LFS server to `gotham_ext`, it seemed like the `lfs_protocol` crate was the logical place for the Git LFS MIME type constant to live. In addition to moving it, this diff swaps out the deprecated `lazy_static` crate with `once_cell`.

Reviewed By: krallin

Differential Revision: D21193938

fbshipit-source-id: 81dc23e8f37a6c0a45ae44443807e5e21214bcd5