tweaks to update the integration tests to allow running on Windows
Summary:
Previously none of the integration tests ran on Windows, as they checked for
the presence of `/dev/fuse` first, and would skip the test if it was not
present.

This updates the tests to allow running on Windows, and updates the
`find_executables.py` code to find the executables properly on Windows.

Reviewed By: wez

Differential Revision: D17733621

fbshipit-source-id: a8bf4e5138f9ea6c9b040814dd68fc64228961de